requirements.txt
begins
pexpect
PyYAML